Repackage launcher and split up interfaces
