+++ UnchainedLauncher.Core/Processes/Chivalry/Chivalry2ServerProcessLauncher.cs
@@ -1,19 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UnchainedLauncher.Core.JsonModels;
-using UnchainedLauncher.Core.Processes;
+﻿using System.Diagnostics;
 
-namespace UnchainedLauncher.Core {
+namespace UnchainedLauncher.Core.Processes.Chivalry {
     public class Chivalry2ServerProcessLauncher {
-        public Chivalry2Launcher Launcher { get; set; }
+        public IUnchainedChivalry2Launcher Launcher { get; set; }
         public ModdedLaunchOptions ModdedLaunchOptions { get; set; }
         public LanguageExt.Option<ServerLaunchOptions> ServerLaunchOptions { get; set; }
         string ExtraArgs { get; set; }
-        public Chivalry2ServerProcessLauncher(Chivalry2Launcher launcher,
+        public Chivalry2ServerProcessLauncher(IUnchainedChivalry2Launcher launcher,
                                            ModdedLaunchOptions options,
                                            string extraArgs) {
             Launcher = launcher;
@@ -22,7 +15,7 @@ namespace UnchainedLauncher.Core {
         }
 
         public LanguageExt.Option<LanguageExt.Either<ProcessLaunchFailure, Process>> Launch() {
-            var launch = Launcher.LaunchUnchained(ModdedLaunchOptions, ExtraArgs);
+            var launch = Launcher.Launch(ModdedLaunchOptions, ExtraArgs);
             return launch;
         }
     }
