+++ UnchainedLauncher.Core/Processes/Chivalry/IOfficialChivalry2Launcher.cs
@@ -1,0 +1,17 @@
+﻿using LanguageExt;
+using System.Diagnostics;
+
+namespace UnchainedLauncher.Core.Processes.Chivalry
+{
+    public interface IOfficialChivalry2Launcher {
+        /// <summary>
+        /// Launches a vanilla game. Implementations may do extra work to enable client side pak file loading
+        /// </summary>
+        /// <param name="args"></param>
+        /// <returns>
+        /// Left if the game failed to launch.
+        /// Right if the game was launched successfully.
+        /// </returns>
+        public Either<ProcessLaunchFailure, Process> Launch(string args);
+    }
+}
